many changes
This is growing into a beast as I think of new ideas and go a little too crazy with them. I've added nearly all functionality from the parameter spreadsheet (minus a full EG section (working on that), MIDI implementation and sequencer) into a number of sub menus that are scrolled through using the face buttons on the control board. I will write more details in the readme.